Add GetLatestNewsCubit and update UI to fetch and display latest news
Implemented the `GetLatestNewsCubit` to handle fetching the latest news and updated the UI components to display the fetched news. This includes changes to `HomeView`, `CustomLatestNewsTile`, `CustomListViewOfLatestNews`, and `HomeViewBody` to integrate the new cubit and display the news dynamically. Additionally, fixed some formatting issues and updated the `main.dart` to include the new cubit in the `MultiBlocProvider`.